Clarify a few areas in the README
Make it more explicit for dependencies that one must both ensure they're listed so they get copied, and they must be included on the pages in order to have effect.

Make note of fact that Hologram also parses javascript files for comments.

Make it clear that items in the same category are sorted alphabetically by name