Issue #86: Add documentation related to categories and parent/child relationships
Make it clear that if a child does not have a category, or defines the same category, it will be grouped with its parent. The implication is that if it doesn't have the same category, it will not appear on the same page.

Make note that the category is optional if a component defines a parent.