Allow template variables to be set as nil
This is needed so that erb files do not have :blocks set sometimes but
nil at other times. Without this change if an erb file is processed
before a documentation page, blocks will (and other related vars) will be
nil while after documentation pages are processed blocks will be set for
erbs. That was unintentional and this fixes it.